+++ crates/sp1-utils/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "sp1-utils"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bincode = { workspace = true }
+sp1-build = { workspace = true }
+sp1-sdk = { workspace = true }
