+++ examples/clob/programs/sp1/Cargo.toml
@@ -7,11 +7,12 @@ edition = "2021"
 path = "src/lib.rs"
 
 [dependencies]
+sp1-utils = { path = "../../../../crates/sp1-utils" }
 once_cell = { workspace = true }
-bincode = { workspace = true }
 
 [build-dependencies]
 sp1-helper = { workspace = true }
 sp1-build = { workspace = true }
 sp1-sdk = { workspace = true }
 bincode = { workspace = true } 
+sp1-utils = { path = "../../../../crates/sp1-utils" }
