+++ examples/matching-game/programs/sp1/src/lib.rs
@@ -1,15 +1,10 @@
 use once_cell::sync::Lazy;
-use std::fs;
+use sp1_utils::get_elf_id;
 
 pub const MATCHING_GAME_ELF: &[u8] = include_bytes!("../program/elf/matching-game-sp1-guest");
 
 static MATCHING_GAME_ELF_ID: Lazy<[u8; 32]> = Lazy::new(|| {
-    let vkey_path =
-        concat!(env!("CARGO_MANIFEST_DIR"), "/program/elf/matching-game-sp1-guest.vkey");
-    let vk_bytes: Vec<u8> =
-        bincode::deserialize_from(fs::File::open(vkey_path).expect("Failed to open vkey file"))
-            .expect("Failed to deserialize verifying key");
-    vk_bytes.try_into().expect("Invalid verifying key length")
+    get_elf_id(concat!(env!("CARGO_MANIFEST_DIR"), "/program/elf/matching-game-sp1-guest.vkey"))
 });
 
 pub fn get_matching_game_program_id() -> [u8; 32] {
